+++ NEWS.md
@@ -1,3 +1,9 @@
+# wgsaparsr 6.2.4
+
+* added outputName as a field to config file (not using it yet, though)
+* changed dplyr::tibble() to tibble::tibble() to avoid warning (mostly in tests)
+* updated DESCRIPTION to import tibble library
+
 # wgsaparsr 6.2.3
 
 * Added a `NEWS.md` file to track changes to the package.